Refactor resetEmailHandler
Move the email part out to its own file so we can use it more easily in the incident report handler.
